# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PRC API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "api/"
        }
    ],
    "paths": {
...
        "/team-list": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Get All Teams",
                "description": "Get team list",
                "operationId": "getTeams",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Team list have been retrieve successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/team": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Team Detail",
                "description": "Get Team details based on team_id",
                "operationId": "getTeam",
                "parameters": [
                    {
                        "name": "team_id",
                        "in": "query",
                        "description": "Pass team id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Team detail has been retrieve successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/add-manager": {
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Add Manager in team",
                "description": "Add manager in team",
                "operationId": "addManager",
                "requestBody": {
                    "description": "Pass Manager details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "User"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "profile_picture": {
                                        "type": "string",
                                        "example": "base64 string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Manager has been added successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/edit-team-member": {
            "put": {
                "tags": [
                    "Teams"
                ],
                "summary": "Edit member of team",
                "description": "Edit member of team",
                "operationId": "editTeamMember",
                "requestBody": {
                    "description": "Pass Member details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member_id",
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "member_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "User"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "PRC"
                                    },
                                    "profile_picture": {
                                        "type": "string",
                                        "example": "Base64 string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Team member update successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/remove-team-member": {
            "delete": {
                "tags": [
                    "Teams"
                ],
                "summary": "team member delete",
                "description": "remove team member",
                "operationId": "removeTeamMember",
                "parameters": [
                    {
                        "name": "member_id",
                        "in": "query",
                        "description": "member id",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The member has been removed from the team."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/add-coach": {
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Add Coach in team",
                "description": "Add coach in team",
                "operationId": "addCoach",
                "requestBody": {
                    "description": "Pass Coach details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "User"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "profile_picture": {
                                        "type": "string",
                                        "example": "base64 string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Coach has been added successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/add-player": {
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Add Player in team",
                "description": "Add player in team",
                "operationId": "addPlayer",
                "requestBody": {
                    "description": "Pass Player details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "User"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "profile_picture": {
                                        "type": "string",
                                        "example": "base64 string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Player has been added successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "apiKey",
                "description": "Authentication Token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}